model_requirements.txt
--find-links https://download.pytorch.org/whl/torch_stable.html

h2o==3.32.0.4
Keras==2.7.0
lightgbm==2.2.2
matplotlib==3.3.4
numba==0.53.1
numexpr==2.7.2
onnx==1.9.0
pamela==1.0.0
patsy==0.5.1
plotly==4.14.3
py4j==0.10.9.1
pyaml==20.4.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycurl==7.43.0.6
pynndescent==0.5.2
pypmml==0.9.9
pytorch-ignite==0.4.3
scikit-learn==1.1.3
scikit-optimize==0.8.1
scipy==1.7.0
seaborn==0.11.1
statsmodels==0.12.2
tensorflow==2.7.1
torch==1.7.1
umap-learn==0.4.1
xgboost==0.81
idna==3.4
pyparsing==2.4.7
pillow==9.0.1
six==1.15.0
certifi==2022.12.7
requests==2.26.0
pyyaml==5.4.1
numpy==1.22.0
aloe==0.2.0
astroid==2.5.1
cloudpickle==1.6.0
livereload==2.6.3
lunr==0.5.8
nltk==3.6.6
pandas==1.3.0
pytest==6.2.2
babel==2.10.3
typed-ast==1.4.2
aiogqlc==2.2.0
validators==0.18.2
coloredlogs==15.0.1
aiohttp==3.8.1
charset-normalizer==2.0.12
nest-asyncio==1.5.4
click==8.0.3
jinja2==3.0.3
markupsafe==2.1.1
packaging==20.9
pytz==2022.7.1
urllib3==1.26.6
pluggy==0.13.1
protobuf==3.20.3
typing-extensions==4.6.0
attrs==22.2.0
cachetools==5.2.0
google-auth==2.15.0
multidict==6.0.3
tornado==6.1
pygments==2.16.1
autodocsumm>=0.2.0
cachecontrol[filecache]==0.12.6
docutils==0.17.1
msgpack==1.0.2
yarl==1.8.2
ruamel-yaml==0.17.22
cssutils==2.6.0
platformdirs==3.7.0
frozenlist==1.3.3
tensorflow-io-gcs-filesystem==0.32.0
sphinx==3.5.1
sphinx-autoapi==1.7.0
sphinx-autodoc-typehints==1.11.1
sphinx-jinja2-compat==0.2.0
sphinx-prompt==1.1.0
sphinx-tabs==3.3.1
sphinx-toolbox==3.0.0
sphinxcontrib-applehelp==1.0.4
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
async-timeout==4.0.2
soupsieve==2.4.1
cycler==0.12.1
executing==1.2.0
fastparquet==0.8.3